UltraBlame original commit: 8da6b16bdd33a7c882d9aa04c05277c911c74fae

+++ layout/base/nsCSSFrameConstructor.cpp
@@ -4715,7 +4715,7 @@ nsCSSFrameConstructor::FindSVGData(Element* aElement,
   if (parentContent) {
     PRInt32 parentNSID;
     nsIAtom* parentTag =
-      parentContent->GetOwnerDoc()->BindingManager()->
+      parentContent->OwnerDoc()->BindingManager()->
         ResolveTag(aParentFrame->GetContent(), &parentNSID);
 
     
@@ -9289,8 +9289,14 @@ nsCSSFrameConstructor::CreateNeededTablePseudos(nsFrameConstructorState& aState,
           bool trailingSpaces = spaceEndIter.SkipWhitespace(aState);
 
           
-          if (trailingSpaces ||
-              spaceEndIter.item().DesiredParentType() != eTypeBlock) {
+          
+          
+          
+          
+          
+          if ((trailingSpaces && ourParentType != eTypeBlock) ||
+              (!trailingSpaces && spaceEndIter.item().DesiredParentType() != 
+               eTypeBlock)) {
             bool updateStart = (iter == endIter);
             endIter.DeleteItemsTo(spaceEndIter);
             NS_ASSERTION(trailingSpaces == endIter.IsDone(), "These should match");
